Update sin of angle in rad .py

+++ sin of angle in rad .py
@@ -1,4 +1,4 @@
 import math
 A= int(raw_input())
 r= math.radians(A)
-print(round(math.sin(r),3))
+print(int(round(math.sin(r),3)))
